+++ src/AzDocs.Common/public/DiagnosticSettings-Helper-Functions.ps1
@@ -5,11 +5,23 @@ function Set-DiagnosticSettings
         [Parameter(Mandatory)][string] $ResourceId,
         [Parameter(Mandatory)][string] $ResourceName, 
         [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceResourceId,
-        [Parameter()][string] $Metrics,
-        [Parameter()][string] $Logs
+        [Parameter()][System.Object[]] $DiagnosticSettingsLogs,
+        [Parameter()][System.Object[]] $DiagnosticSettingsMetrics
     )
 
     Write-Header -ScopedPSCmdlet $PSCmdlet
+
+    if (!$DiagnosticSettingsLogs)
+    {
+        Write-Host "No diagnostic settings for logs were specified. Continueing with the default set"
+        $DiagnosticSettingsLogs = Get-DefaultDiagnosticSettings -ResourceId $ResourceId -DiagnosticSettingType 'Logs'
+    }
+
+    if (!$DiagnosticSettingsMetrics)
+    {
+        Write-Host "No diagnostic settings for metrics were specified. Continueing with the default set"
+        $DiagnosticSettingsMetrics = Get-DefaultDiagnosticSettings -ResourceId $ResourceId -DiagnosticSettingType 'Metrics'
+    }
 
     # Get root path and make sure the right provider is registered
     $RootPath = Split-Path (Split-Path $PSScriptRoot -Parent) -Parent
@@ -18,67 +30,127 @@ function Set-DiagnosticSettings
     # Set the name for the diagnostic setting 
     $diagnosticSettingName = "$ResourceName-diagnostic-setting"
 
+    # Remove existing diagnostic setting
+    Remove-DiagnosticSetting -ResourceId $ResourceId -LogAnalyticsWorkspaceResourceId $LogAnalyticsWorkspaceResourceId -ResourceName $ResourceName
+
+    # Create the diagnostic setting
     $optionalParameters = @()
-    if ($Metrics)
+    if ($DiagnosticSettingsMetrics)
     {
-        $optionalParameters += "--metrics", "$Metrics"
-    }
-    if ($Logs)
-    {
-        $optionalParameters += "--logs", "$Logs"
-    }
+        # Confirm the diagnostic setting exists
+        Confirm-DiagnosticSettings -ResourceId $ResourceId -DiagnosticSettings $DiagnosticSettingsMetrics -DiagnosticSettingType "Metrics"
 
-    # For backwards compatibility, remove the old diagnostic settings 
-    $oldDiagnosticSettings = Invoke-Executable az monitor diagnostic-settings list --resource $ResourceId | ConvertFrom-Json
-    foreach ($oldDiagnosticSetting in $oldDiagnosticSettings.value)
-    {
-        if ($oldDiagnosticSetting.name -and $oldDiagnosticSetting.name -ne $diagnosticSettingName -and $oldDiagnosticSetting.workspaceId -eq $LogAnalyticsWorkspaceResourceId)
-        {
-            # remove old diagnostic setting
-            Invoke-Executable az monitor diagnostic-settings delete --name $oldDiagnosticSetting.name --resource $ResourceId
-        }
+        $metrics = New-DiagnosticSetting -DiagnosticSettings $DiagnosticSettingsMetrics
+        $optionalParameters += "--metrics", "$metrics"
     }
+    if ($DiagnosticSettingsLogs)
+    {
+        # Confirm the diagnostic setting exists
+        Confirm-DiagnosticSettings -ResourceId $ResourceId -DiagnosticSettings $DiagnosticSettingsLogs -DiagnosticSettingType "Logs"
 
+        $logs = New-DiagnosticSetting -DiagnosticSettings $DiagnosticSettingsLogs
+        $optionalParameters += "--logs", "$logs"
+    }
+       
     # Create new diagnostic setting
     Invoke-Executable az monitor diagnostic-settings create --resource $ResourceId --name $diagnosticSettingName --workspace $LogAnalyticsWorkspaceResourceId @optionalParameters
 
     Write-Footer -ScopedPSCmdlet $PSCmdlet
 }
 
-function Get-DiagnosticSettingBasedOnTier
+function Remove-DiagnosticSetting
 {
     [CmdletBinding()]
     param (
-        [Parameter(Mandatory)][ValidateSet("webapp")][string] $ResourceType,
-        [Parameter(Mandatory)][string] $CurrentResourceTier
+        [Parameter(Mandatory)][string] $ResourceId,
+        [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceResourceId,
+        [Parameter(Mandatory)][string] $ResourceName
     )
-
-    #  Create diagnostics settings & make sure to check which categories can be accessed based upon the SKU of the app service plan
-    # https://docs.microsoft.com/en-us/azure/app-service/troubleshoot-diagnostic-logs#send-logs-to-azure-monitor-preview
 
     Write-Header -ScopedPSCmdlet $PSCmdlet
 
-    $diagnosticSettingToReturn = $null
-    if ($ResourceType -eq 'webapp')
+    $diagnosticSettingName = "$ResourceName-diagnostic-setting"
+
+    $existingDiagnosticSettings = Invoke-Executable az monitor diagnostic-settings list --resource $ResourceId | ConvertFrom-Json
+    foreach ($diagnosticSetting in $existingDiagnosticSettings.value)
     {
-        $allowedWebAppTiers = @("Premium", "PremiumV2", "PremiumV3")
-        if ($allowedWebAppTiers -contains $CurrentResourceTier)
+        if ($diagnosticSetting.name -and $diagnosticSetting.name -eq $DiagnosticSettingName -and $diagnosticSetting.workspaceId -eq $LogAnalyticsWorkspaceResourceId)
         {
-            $diagnosticSettingToReturn = [PSCustomObject]@{
-                DiagnosticSettingType  = "Logs"
-                DiagnosticSettingValue = "[{ 'category': 'AppServiceHTTPLogs', 'enabled': true }, { 'category': 'AppServiceConsoleLogs', 'enabled': true }, { 'category': 'AppServiceAppLogs', 'enabled': true }, { 'category': 'AppServiceFileAuditLogs', 'enabled': true }, { 'category': 'AppServiceIPSecAuditLogs', 'enabled': true }, { 'category': 'AppServicePlatformLogs', 'enabled': true }, { 'category': 'AppServiceAuditLogs', 'enabled': true } ]".Replace("'", '\"')
-            }
+            Write-Host "Removing diagnostic settings $DiagnosticSettingName"
+            # remove old diagnostic setting
+            Invoke-Executable az monitor diagnostic-settings delete --name $diagnosticSetting.name --resource $ResourceId
         }
-        else
+    }
+
+    Write-Footer -ScopedPSCmdlet $PSCmdlet
+}
+
+function Get-DefaultDiagnosticSettings
+{
+    [CmdletBinding()]
+    param (
+        [Parameter()][string] $ResourceId,
+        [Parameter()][string][ValidateSet("Logs", "Metrics")] $DiagnosticSettingType
+    )
+
+    Write-Header -ScopedPSCmdlet $PSCmdlet
+
+    $availableDiagnosticSettings = (Invoke-Executable az monitor diagnostic-settings categories list --resource $ResourceId | ConvertFrom-Json).value
+    $settings = ($availableDiagnosticSettings | Where-Object { $_.categoryType -eq $DiagnosticSettingType }).name
+    
+    Write-Output $settings
+    Write-Footer -ScopedPSCmdlet $PSCmdlet
+}
+
+function New-DiagnosticSetting
+{
+    [CmdletBinding()]
+    param (
+        [Parameter()][System.Object[]] $DiagnosticSettings
+    )
+
+    Write-Header -ScopedPSCmdlet $PSCmdlet
+
+    $diagnosticSettingsToCreate = @()
+    foreach ($diagnosticSetting in $DiagnosticSettings)
+    {
+        $diagnosticSettingsToCreate += [PSCustomObject]@{
+            Category = $diagnosticSetting
+            Enabled  = $true
+        }
+    }
+
+    Write-Output ($diagnosticSettingsToCreate | ConvertTo-Json -Compress -AsArray).Replace('"', '\"')
+    Write-Footer -ScopedPSCmdlet $PSCmdlet
+}
+
+function Confirm-DiagnosticSettings
+{
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)][string] $ResourceId,
+        [Parameter()][System.Object[]] $DiagnosticSettings,
+        [Parameter()][string][ValidateSet("Logs", "Metrics")] $DiagnosticSettingType
+    )
+
+    Write-Header -ScopedPSCmdlet $PSCmdlet
+
+    $availableDiagnosticSettings = (Invoke-Executable az monitor diagnostic-settings categories list --resource $ResourceId | ConvertFrom-Json).value
+    Write-Host "Available diagnostic settings $($availableDiagnosticSettings.name)"
+    if ($DiagnosticSettings)
+    {
+        foreach ($diagnosticSetting in $DiagnosticSettings) 
         {
-            $diagnosticSettingToReturn = [PSCustomObject]@{
-                DiagnosticSettingType  = "Logs"
-                DiagnosticSettingValue = "[{ 'category': 'AppServiceHTTPLogs', 'enabled': true }, { 'category': 'AppServiceConsoleLogs', 'enabled': true }, { 'category': 'AppServiceAppLogs', 'enabled': true }, { 'category': 'AppServiceIPSecAuditLogs', 'enabled': true }, { 'category': 'AppServicePlatformLogs', 'enabled': true }, { 'category': 'AppServiceAuditLogs', 'enabled': true } ]".Replace("'", '\"')
+            $foundDiagnosticSettings = $availableDiagnosticSettings | Where-Object { $_.categoryType -eq $DiagnosticSettingType }
+            if (!($foundDiagnosticSettings.name -contains $diagnosticSetting))
+            {
+                Write-Output $false
+                throw "Diagnostic setting $DiagnosticSettingType category: $diagnosticSetting  is unknown for this resource. Please specify a category that exists."
             }
         }
     }
 
-    Write-Output $diagnosticSettingToReturn
+    Write-Output $true
 
     Write-Footer -ScopedPSCmdlet $PSCmdlet
 }